backend/requirements.txt
asgiref==3.8.1
Django==5.0.9
django-cors-headers==4.6.0
django-filter==24.3
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
mssql-django==1.5
packaging==24.2
pillow==11.0.0
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1